listener: Pass an initialized Listener to plugins
MutableObjects.Listener must be initialized prior to calling plugins:

    type MutableObjects struct {
            // Listener is the listener being built. Must be initialized before Plugin methods are called.
            Listener *listener.Listener
    ...
    }

This is needed for the plugins to be able to add listener filters. Pick
up any listener filters added by plugins in buildInboundFilterchains().

Outbound code creates the listener prior to calling the plugins, which
can then add listener filters to the final listener directly.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>